Mesh: fix two bugs with Mesh.append; see #19
appendData was incorrectly returning reflect.Value when we wanted
it to return an interface, this causes a crash because callers expect
to derive the underlying type from the interface value.

fixIndices was sometimes not intialized (in the case of appending a
non-indexed mesh to another non-indexed mesh), which caused a nil
pointer panic.